Added progress bar indicators for bookmarks
Add progress bar indicators that turn into buttons on hover. Adds good asthetic to the progress bar making it more functional. visuals are good for people to not have to keep track of bookmarks.

Added multiple todos for storage and reseting the state when a new video is clicked as well as loading bookmarks using chrome storage