Added sync storage for bookmarks
Added functionaliity to save tabs in chrome storage for videos. Added methods for add, clear, query, rem, queryAll to debug and make functional storage.

Inject.js handles notifying the backgorund script of when it needs to update storage and the background script is what communicates with storage.

Added beforeClose listener to the tab in order to save bookmarks to do one last sync so the storage is the same as the local statae before the tab is closed.